fix warnings / tidy code
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@51301 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ desktop/src/main/java/org/jfugue/MidiParser.java
@@ -220,9 +220,9 @@ public final class MidiParser extends Parser {
 
 		fireTimeEvent(new Time(time));
 		fireVoiceEvent(new Voice((byte) track));
-		Note note = new Note((byte) data1, (long) (timestamp - time));
+		Note note = new Note((byte) data1, timestamp - time);
 		note.setDecimalDuration(
-				(double) ((timestamp - time) / (this.tempo * 4.0D)));
+				(timestamp - time) / (this.tempo * 4.0D));
 		note.setAttackVelocity(tempNoteAttackRegistry[track][data1]);
 		note.setDecayVelocity((byte) data2);
 		fireNoteEvent(note);
